change routing from auth to (auth)
