+++ src/app/(auth)/signup/page.tsx
@@ -3,25 +3,23 @@
 import { useState } from 'react';
 import { useRouter } from 'next/navigation';
 import TextInput from '@/components/TextInput/index';
-import { H1, H4, P, AColored } from '@/styles/text';
+import { H1, H4, AColored } from '@/styles/text';
 import supabase from '@/api/supabase/createClient';
 import COLORS from '@/styles/colors';
 import {
   OuterDiv,
+  SpacerDiv,
   FormDiv,
   H4Centered,
   QuestionsDiv,
-  SpacerDiv,
-} from '@/app/auth/styles';
+} from '@/app/(auth)/styles';
 import BigButton from '@/components/BigButton';
 
-export default function Login() {
+export default function SignUp() {
   const [email, setEmail] = useState('');
   const [password, setPassword] = useState('');
   // const [errorMessage, setErrorMessage] = useState('');
   const { push } = useRouter();
-  // commenting out Sign Up to use in the next PR -> create a separate sign in page
-  /*
   const handleSignUp = async () => {
     const { error } = await supabase.auth.signUp({
       email,
@@ -33,24 +31,11 @@ export default function Login() {
     }
     push('/');
   };
-  */
-
-  const handleSignIn = async () => {
-    const { error } = await supabase.auth.signInWithPassword({
-      email,
-      password,
-    });
-
-    if (error) {
-      throw new Error(`An error occurred trying to sign in: ${error.message}`);
-    }
-    push('/');
-  };
 
   return (
     <OuterDiv>
       <FormDiv>
-        <H1>Log In</H1>
+        <H1>Sign Up</H1>
         <SpacerDiv>
           <QuestionsDiv>
             <TextInput
@@ -74,20 +59,15 @@ export default function Login() {
               setValue={setPassword}
             />
           </QuestionsDiv>
-          <P>
-            <AColored $color={COLORS.greyMid} href="/reset-password">
-              Forgot your password?
-            </AColored>
-          </P>
         </SpacerDiv>
         <SpacerDiv>
-          <BigButton type="button" onClick={handleSignIn}>
-            <H4 $color="white">Sign in</H4>
+          <BigButton type="button" onClick={handleSignUp}>
+            <H4 $color="white">Sign Up</H4>
           </BigButton>
           <H4Centered>
-            Don’t have an account yet?{' '}
-            <AColored $color={COLORS.greyDark} href="/">
-              Sign up
+            Have an account already?{' '}
+            <AColored $color={COLORS.greyDark} href="/login">
+              Log In
             </AColored>
           </H4Centered>
         </SpacerDiv>
