+++ src/app/(auth)/styles.tsx
